Put newslist into container, added timestamps
