+++ package.json
@@ -16,6 +16,7 @@
     "@mui/material": "^5.13.4",
     "axios": "^1.4.0",
     "chart.js": "^4.3.0",
+    "date-fns": "^2.30.0",
     "dotenv": "^16.1.4",
     "firebase": "^9.22.2",
     "react": "^18.2.0",
