+++ src/css/App.css
@@ -44,4 +44,8 @@ a:link {
   color: green;
   background-color: transparent;
   text-decoration: none;
+  font-weight: 700;
+}
+a:visited {
+  color: blue;
 }
